changelog: Move markdown rendering code into MarkdownRenderer class
